requirements.txt
pdfminer.six
PyMuPDF
azure-ai-documentintelligence
python-dotenv
